backend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2
django-phonenumber-field[phonenumbers]
Pillow